Backend/requirements.txt
django
google-generativeai
python-dotenv
subprocess.run
openai
django-cors-headers
gitpython
docker